Split up type inference test file
